requirements.txt
discord.py==2.4.0
rcon==2.1.1
